+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/domain/Disableable.java
@@ -15,7 +15,7 @@ public interface Disableable {
 	boolean isDisabled();
 
 	/**
-	 * Enable / diable entity
+	 * Enable / disable entity
 	 * 
 	 * @param disabled
 	 */