+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmIdentityContractService.java
@@ -81,6 +81,13 @@ public class DefaultIdmIdentityContractService
 	}
 	
 	@Override
+	protected IdmIdentityContract toEntity(IdmIdentityContractDto dto, IdmIdentityContract entity) {
+		IdmIdentityContract contract = super.toEntity(dto, entity);
+		contract.setDisabled(dto.isDisabled()); // redundant attribute for queries
+		return contract;
+	}
+	
+	@Override
 	protected List<Predicate> toPredicates(Root<IdmIdentityContract> root, CriteriaQuery<?> query, CriteriaBuilder builder, IdmIdentityContractFilter filter) {
 		List<Predicate> predicates = super.toPredicates(root, query, builder, filter);
 		// quick
@@ -112,41 +119,49 @@ public class DefaultIdmIdentityContractService
 		if (filter.getMain() != null) {
 			predicates.add(builder.equal(root.get(IdmIdentityContract_.main), filter.getMain()));
 		}
-		if (filter.getValid() != null && filter.getValid()) {
-			final LocalDate today = LocalDate.now();
-			predicates.add(
-					builder.and(
-							builder.or(
-									builder.lessThanOrEqualTo(root.get(IdmIdentityContract_.validFrom), today),
-									builder.isNull(root.get(IdmIdentityContract_.validFrom))
-									),
-							builder.or(
-									builder.greaterThanOrEqualTo(root.get(IdmIdentityContract_.validTill), today),
-									builder.isNull(root.get(IdmIdentityContract_.validTill))
-									)
-							)
-					);
-		}
-		if (filter.getValid() != null && !filter.getValid()) {
-			final LocalDate today = LocalDate.now();
-			predicates.add(
-					builder.or(
-							builder.lessThan(root.get(IdmIdentityContract_.validTill), today),
-							builder.greaterThan(root.get(IdmIdentityContract_.validFrom), today)
-							)
-					);
+		if (filter.getValid() != null) {
+			if (filter.getValid()) {
+				final LocalDate today = LocalDate.now();
+				predicates.add(
+						builder.and(
+								builder.or(
+										builder.lessThanOrEqualTo(root.get(IdmIdentityContract_.validFrom), today),
+										builder.isNull(root.get(IdmIdentityContract_.validFrom))
+										),
+								builder.or(
+										builder.greaterThanOrEqualTo(root.get(IdmIdentityContract_.validTill), today),
+										builder.isNull(root.get(IdmIdentityContract_.validTill))
+										),
+								builder.equal(root.get(IdmIdentityContract_.disabled), false)
+								)								
+						);
+			} else {
+				final LocalDate today = LocalDate.now();
+				predicates.add(
+						builder.or(
+								builder.lessThan(root.get(IdmIdentityContract_.validTill), today),
+								builder.greaterThan(root.get(IdmIdentityContract_.validFrom), today),
+								builder.equal(root.get(IdmIdentityContract_.disabled), true)
+								)
+						);
+			}
 		}
 		if (filter.getValidNowOrInFuture() != null) {
 			if (filter.getValidNowOrInFuture()) {
 				predicates.add(
-							builder.or(
-									builder.greaterThanOrEqualTo(root.get(IdmIdentityContract_.validTill), LocalDate.now()),
-									builder.isNull(root.get(IdmIdentityContract_.validTill))
-									)
-						);
+						builder.and(
+								builder.or(
+										builder.greaterThanOrEqualTo(root.get(IdmIdentityContract_.validTill), LocalDate.now()),
+										builder.isNull(root.get(IdmIdentityContract_.validTill))
+										),
+								builder.equal(root.get(IdmIdentityContract_.disabled), false)
+							));
 			} else {
 				predicates.add(builder.lessThan(root.get(IdmIdentityContract_.validTill), LocalDate.now()));
 			}
+		}
+		if (filter.getState() != null) {
+			predicates.add(builder.equal(root.get(IdmIdentityContract_.state), filter.getState()));
 		}
 		
 		return predicates;
@@ -207,7 +222,7 @@ public class DefaultIdmIdentityContractService
 	 * - 2. valid (validable and not disabled)
 	 * - 3. with working position with default tree type
 	 * - 4. with working position with any tree type
-	 * - 5. other
+	 * - 5. other with lowest valid from
 	 * 
 	 * @param identityId
 	 * @return
@@ -272,7 +287,6 @@ public class DefaultIdmIdentityContractService
 			//
 			return builder.toComparison();
 		}
-		
 	}
 }