+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmIdentityContractServiceUnitTest.java
@@ -29,7 +29,8 @@ import eu.bcvsolutions.idm.test.api.AbstractUnitTest;
 
 /**
  * Identity contract unit tests:
- * - 
+ * - prime contract
+ * - contract state
  * 
  * @author Radek Tomiška
  */